backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib
python-jose[cryptography]
python-dotenv
email-validator
python-multipart
